requirements.txt
langchain
langchain-community
langchain-ollama
pillow